+++ responses.go
@@ -94,32 +94,34 @@ type ProcessedTransaction struct {
 
 type ActionTraceReceipt struct {
 	Receiver        AccountName                    `json:"receiver"`
-	ActionDigest    string                         `json:"act_digest"`
+	ActionDigest    Checksum256                    `json:"act_digest"`
 	GlobalSequence  Uint64                         `json:"global_sequence"`
 	ReceiveSequence Uint64                         `json:"recv_sequence"`
 	AuthSequence    []TransactionTraceAuthSequence `json:"auth_sequence"` // [["account", sequence], ["account", sequence]]
-	CodeSequence    Uint64                         `json:"code_sequence"`
-	ABISequence     Uint64                         `json:"abi_sequence"`
+	CodeSequence    Varuint32                      `json:"code_sequence"`
+	ABISequence     Varuint32                      `json:"abi_sequence"`
 }
 
 type ActionTrace struct {
-	Receipt                                *ActionTraceReceipt `json:"receipt,omitempty"`
+	ActionOrdinal                          Varuint32           `json:"action_ordinal"`
+	CreatorActionOrdinal                   Varuint32           `json:"creator_action_ordinal"`
+	ClosestUnnotifiedAncestorActionOrdinal Varuint32           `json:"closest_unnotified_ancestor_action_ordinal"`
+	Receipt                                *ActionTraceReceipt `json:"receipt,omitempty" eos:"optional"`
 	Receiver                               AccountName         `json:"receiver"`
 	Action                                 *Action             `json:"act"`
+	ContextFree                            bool                `json:"context_free"`
 	Elapsed                                Int64               `json:"elapsed"`
 	Console                                string              `json:"console"`
 	TransactionID                          Checksum256         `json:"trx_id"`
-	InlineTraces                           []ActionTrace       `json:"inline_traces"`
-	ContextFree                            bool                `json:"context_free"`
-	BlockTime                              BlockTimestamp      `json:"block_time"`
 	BlockNum                               uint32              `json:"block_num"`
-	ProducerBlockID                        Checksum256         `json:"producer_block_id"`
-	AccountRAMDeltas                       []*AccountRAMDelta  `json:"account_ram_deltas"`
-	ErrorCode                              *Uint64             `json:"error_code"`
-	Except                                 *Except             `json:"except"`
-	ActionOrdinal                          uint32              `json:"action_ordinal"`
-	CreatorActionOrdinal                   uint32              `json:"creator_action_ordinal"`
-	ClosestUnnotifiedAncestorActionOrdinal uint32              `json:"closest_unnotified_ancestor_action_ordinal"`
+	BlockTime                              BlockTimestamp      `json:"block_time"`
+	ProducerBlockID                        Checksum256         `json:"producer_block_id" eos:"optional"`
+	AccountRAMDeltas                       []*AccountRAMDelta  `json:"account_ram_deltas,omitempty"`
+	Except                                 *Except             `json:"except,omitempty" eos:"optional"`
+	ErrorCode                              *Uint64             `json:"error_code,omitempty" eos:"optional"`
+
+	// Not present in EOSIO >= 1.8.x
+	InlineTraces []ActionTrace `json:"inline_traces,omitempty" eos:"-"`
 }
 
 type AccountRAMDelta struct {
@@ -455,7 +457,7 @@ type GetCurrencyStatsResp struct {
 }
 
 type Except struct {
-	Code    int                 `json:"code"`
+	Code    Int64               `json:"code"`
 	Name    string              `json:"name"`
 	Message string              `json:"message"`
 	Stack   []*ExceptLogMessage `json:"stack"`
@@ -468,13 +470,167 @@ type ExceptLogMessage struct {
 	Data    json.RawMessage  `json:"data"`
 }
 
+var exceptLogMessageTypes = map[string]fcVariantType{
+	"context": fcVariantObjectType,
+	"format":  fcVariantStringType,
+}
+
+func (m *ExceptLogMessage) UnmarshalBinary(decoder *Decoder) error {
+	variant := fcVariant{}
+	err := decoder.Decode(&variant)
+	if err != nil {
+		return fmt.Errorf("unable to decode except log message: %s", err)
+	}
+
+	if variant.TypeID != fcVariantObjectType {
+		return fmt.Errorf("invalid log message, expected type %s, got %s", fcVariantObjectType, variant.TypeID)
+	}
+
+	object := variant.MustAsObject()
+
+	if err := object.validateFields(exceptLogMessageTypes); err != nil {
+		return fmt.Errorf("invalid log message object: %s", err)
+	}
+
+	if err = m.Context.fromObject(object["context"].MustAsObject()); err != nil {
+		return fmt.Errorf("unable to assign context: %s", err)
+	}
+
+	m.Format = object["format"].MustAsString()
+
+	if dataVariant := object["data"]; !dataVariant.IsNil() {
+		if m.Data, err = json.Marshal(dataVariant.ToNative()); err != nil {
+			return fmt.Errorf("unable to assign data: %s", err)
+		}
+	}
+
+	return nil
+}
+
 type ExceptLogContext struct {
-	Level      string            `json:"level"` // "debug", "info", "warn", "error", also: "all", "off"
+	Level      ExceptLogLevel    `json:"level"`
 	File       string            `json:"file"`
-	Line       int               `json:"line"`
+	Line       uint64            `json:"line"`
 	Method     string            `json:"method"`
 	Hostname   string            `json:"hostname"`
 	ThreadName string            `json:"thread_name"`
 	Timestamp  JSONTime          `json:"timestamp"`
 	Context    *ExceptLogContext `json:"context,omitempty"`
+}
+
+var exceptLogContextTypes = map[string]fcVariantType{
+	"level":       fcVariantStringType,
+	"file":        fcVariantStringType,
+	"line":        fcVariantUint64Type,
+	"method":      fcVariantStringType,
+	"hostname":    fcVariantStringType,
+	"thread_name": fcVariantStringType,
+	"timestamp":   fcVariantStringType,
+	"?context":    fcVariantObjectType,
+}
+
+func (c *ExceptLogContext) fromObject(object fcVariantObject) error {
+	if err := object.validateFields(exceptLogContextTypes); err != nil {
+		return fmt.Errorf("invalid log context: %s", err)
+	}
+
+	c.Level.FromString(object["level"].MustAsString())
+	c.File = object["file"].MustAsString()
+	c.Line = object["line"].MustAsUint64()
+	c.Method = object["method"].MustAsString()
+	c.Hostname = object["hostname"].MustAsString()
+	c.ThreadName = object["thread_name"].MustAsString()
+
+	var err error
+	if c.Timestamp, err = ParseJSONTime(object["timestamp"].MustAsString()); err != nil {
+		return fmt.Errorf("invalid log context timestamp: %s", err)
+	}
+
+	contextVariant := object["context"]
+	if contextVariant.TypeID != fcVariantNullType {
+		c.Context = new(ExceptLogContext)
+		if err := c.Context.fromObject(contextVariant.MustAsObject()); err != nil {
+			return fmt.Errorf("unable to assign nested context: %s", err)
+		}
+	}
+
+	return nil
+}
+
+type ExceptLogLevel uint8
+
+const (
+	ExceptLogLevelAll ExceptLogLevel = iota
+	ExceptLogLevelDebug
+	ExceptLogLevelInfo
+	ExceptLogLevelWarn
+	ExceptLogLevelError
+	ExceptLogLevelOff
+)
+
+func (s *ExceptLogLevel) FromString(input string) {
+	switch input {
+	case "all":
+		*s = ExceptLogLevelAll
+	case "debug":
+		*s = ExceptLogLevelDebug
+	case "info":
+		*s = ExceptLogLevelInfo
+	case "warn":
+		*s = ExceptLogLevelWarn
+	case "error":
+		*s = ExceptLogLevelError
+	case "off":
+		*s = ExceptLogLevelOff
+	default:
+		*s = ExceptLogLevelOff
+	}
+}
+
+func (s *ExceptLogLevel) UnmarshalJSON(data []byte) error {
+	var decoded string
+	if err := json.Unmarshal(data, &decoded); err != nil {
+		return err
+	}
+
+	s.FromString(decoded)
+	return nil
+}
+
+func (s ExceptLogLevel) MarshalJSON() (data []byte, err error) {
+	out := "off"
+	switch s {
+	case ExceptLogLevelAll:
+		out = "all"
+	case ExceptLogLevelDebug:
+		out = "debug"
+	case ExceptLogLevelInfo:
+		out = "info"
+	case ExceptLogLevelWarn:
+		out = "warn"
+	case ExceptLogLevelError:
+		out = "error"
+	case ExceptLogLevelOff:
+		out = "off"
+	}
+	return json.Marshal(out)
+}
+
+func (s ExceptLogLevel) String() string {
+	switch s {
+	case ExceptLogLevelAll:
+		return "all"
+	case ExceptLogLevelDebug:
+		return "debug"
+	case ExceptLogLevelInfo:
+		return "info"
+	case ExceptLogLevelWarn:
+		return "warn"
+	case ExceptLogLevelError:
+		return "error"
+	case ExceptLogLevelOff:
+		return "off"
+	}
+
+	return "off"
 }