+++ CHANGELOG.md
@@ -6,12 +6,43 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Ability to decode nested `arrays`.
+- Ability to decode nested `arrays` in ABI decoder.
+- Added `BlockState.Header` field of type `SignedBlockHeader` that was previously missing from the struct definition.
+- Added `BlockState.AdditionalSignatures` field of type `[]ecc.Signature` that was previously missing from the struct definition.
+- Added `ActionTrace.ContextFree` field of type `bool` that was previously missing from the struct definition.
+
+### Fixed
+- Unpacking binary `Except` now correctly works.
+- Unpacking binary `Action` and `ActionTrace` now correctly works.
+- Unpacking binary `TransactionTrace` now correctly works.
+- Unpacking binary `TransactionReceipt` type will now correctly set the inner `TransactionWithID.ID` field correctly.
+- Unpacking binary `BlockState` now correctly works but is restricted to EOSIO 2.0.x version.
 
 ### Changed
+- **BREAKING**: Fixed binary unpacking of `BlockState`, `TransactionTrace`, `SignedTransaction`, `Action` (and some inner types). This required changing a few struct fields to better fit with EOSIO definition, here the full list:
+  - `MerkleRoot.ActiveNodes` is now a `[]Checksum256`, was previously `[]string`
+  - `MerkleRoot.NodeCount` is now a `uint64`, was previously `uint32`
+  - Type `EOSNameOrUint32` has been removed and replaced by `PairAccountNameBlockNum` which is strictly typed now.
+  - `BlockState.ProducerToLastProduced` is now `[]PairAccountNameBlockNum`, was previously `[][2]EOSNameOrUint32`.
+  - `BlockState.ProducerToLastImpliedIRB` is now `[]PairAccountNameBlockNum`, was previously `[][2]EOSNameOrUint32`.
+  - `BlockState.BlockID` is now a `Checksum256`, was previously `string`.
+  - `BlockState.ActivatedProtocolFeatures` is now a `*ProtocolFeatureActivationSet`, was previously `map[string][]HexBytes`.
+  - `PendingSchedule.ScheduleHash` is now a `Checksum256`, was previously `HexBytes`.
+  - `ActionTraceReceipt.ActionDigest` is now a `Checksum256`, was previously `string`.
+  - `ActionTraceReceipt.CodeSequence` is now a `Varuint32`, was previously `Uint64`.
+  - `ActionTraceReceipt.ABISequence` is now a `Varuint32`, was previously `Uint64`.
+  - `ActionTrace.ActionOrdinal` is now a `Varuint32`, was previously `uint32`.
+  - `ActionTrace.CreatorActionOrdinal` is now a `Varuint32`, was previously `uint32`.
+  - `ActionTrace.ClosestUnnotifiedAncestorActionOrdinal` is now a `Varuint32`, was previously `uint32`.
+  - `Except.Code` is now a `Int64`, was previously `int`.
+  - `ExceptLogContext.Level` is now a `ExceptLogLevel`, was previously `string`.
+  - `ExceptLogContext.Line` is now a `uint64`, was previously `int`.
+
+  **Note** While those are flagged as breaking change to augment the visibility, they are really bug fixes to fit with the behavior of `nodeos` directly.
+
 - **BREAKING**: The decoding for ABI `variant` was not returning the correct `json` representation. Now ABI `variant` is returned as a two elements array, the first element being the `variant` type name as a `string` and the second the actual value as JSON. For example, assuming a `variant` type defined as `game_type: ["string", "uint32"]`, and a `field` of type `game_type`, before, the JSON serialization would have looked like `{"field":"some_string"}` or `{"field":100}` while after the change, it will get serialized to the correct form `{"field":["string", "some_string"]}` or `{"field":["uint32", 100]}`.
 
-  **Note** While this is flagged a breaking change to augment the visibility, this is really a bug fix to fit with the behavior of `nodeos` directly.
+  **Note** While this is flagged as breaking change to augment the visibility, this is really a bug fix to fit with the behavior of `nodeos` directly.
 
 - **BREAKING**: The serialization for `ExtendedAsset` was aligned with the `eos` codebase.  Beforehand, it would serialize the field name `"Contract"` with a capital `C`, and the `Asset` field as `"asset"` where it should have been `"quantity"`.